requirements.txt
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
av==14.4.0
cdifflib==1.2.9
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
comm==0.2.2
cuda-bindings==13.2.0
cuda-pathfinder==1.5.2
cuda-toolkit==13.0.2
Cython==3.1.1
debugpy==1.8.14
decorator==5.2.1
editdistance==0.8.1
einops==0.8.2
einx==0.3.0
encodec==0.1.1
exceptiongroup==1.3.0
executing==2.2.0
faiss-cpu==1.13.2
fastapi==0.135.3
ffmpy==0.5.0
filelock==3.25.2
frozendict==2.4.6
fsspec==2026.3.0
gradio==5.31.0
gradio_client==1.10.1
groovy==0.1.2
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
importlib_resources==6.5.2
inflect==7.5.0
ipykernel==6.29.5
ipython==8.36.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.3
jupyter_client==8.6.3
jupyter_core==5.7.2
libcst==1.8.6
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==11.0.2
mpmath==1.3.0
nemo_text_processing==1.0.2
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
nvidia-cublas==13.1.0.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti==13.0.85
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime==13.0.96
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver==12.0.4.66
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse==12.6.3.3
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu12==2.21.5
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
nvidia-nvtx-cu12==12.1.105
orjson==3.10.18
packaging==26.0
pandas==2.3.3
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
pip==26.0.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pybase16384==0.3.8
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
pydub==0.25.1
Pygments==2.19.1
pynini==2.1.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq==26.4.0
regex==2026.4.4
requests==2.33.1
rich==14.0.0
ruff==0.11.11
sacremoses==0.1.1
safehttpx==0.1.6
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
semantic-version==2.10.0
sentence-transformers==5.4.0
setuptools==70.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
stack-data==0.6.3
starlette==0.52.1
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.19.1
tomlkit==0.13.3
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.5.1
tqdm==4.67.3
traitlets==5.14.3
transformers==4.41.0
triton==3.1.0
typeguard==4.4.2
typer==0.24.1
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2026.1
urllib3==2.6.3
uv==0.7.3
uvicorn==0.34.2
vector-quantize-pytorch==1.22.16
vocos==0.1.0
wcwidth==0.2.13
websockets==15.0.1
WeTextProcessing==1.0.3
wget==3.2
wheel==0.46.3
wrapt==1.17.2
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.0
numpy>=1.24.0
aiohttp>=3.9.0
python-multipart>=0.0.6